Disable the "Send Private Message" buttons when the signature fails.
If there is no valid signature, we can't be sure that the certificate
belongs to the person who wrote the message. Hence we disable the option
to send private messages.

The application needs to tag the message with a new attribute 'ecca_id',
it must contain a value that's unique on the current page. Then it needs
to add an attribute 'ecca_id_ref' with the same value as the message's
ecca_id to each element that needs to be disabled when verification
fails.

Disabling is done by removing any href-attributes and by adding
'ecca_disabled' to the element's class attribute. The application
designer could use that to change the appearance.