Added Direct Connection functionality.
With it, a user can invite another user to connect directly, i.e. without going through an
intermediate party.

The inviter creates a listening socket, send the ip and port numbers to the invitee and awaits
an incoming connection.

The invitee receives a encrypted and signed message from the invitee. Our proxy agent presents
a button that connects to the inviter.

At the moment, we only offer an echo service but who knows what the future brings...