Keep onion sockets working when restarting the proxy service.
When restarting the proxy, we created new socket endpoints where the
tor onion hidden service delivers the connection requests get delivered.
But tor doesn't like that. It expects the same local endpoint. It makes
sense from a hidden web service on an well known port. But not for
ephemeral ports. These might be taken when restarting the proxy at a later
date, preventing that connection from being used.

For now, we reopen the old endpoint, making tor work at a small chance of
breaking (some) connections.