Split templates and use go rice to serve those
Go rice has the advantage of making it possible to embed the external
files into the binary (see the `release` target in the makefile).

It also always loads from disk in a `dev` build, so no proxy
compilations are necessary while changing the external files (i.e. the templates); a page reload will do.